requirements/requirements_test.txt
pytest>=7.2.0
pytest-cov>=4.0.0
pytest-sanic>=1.9.1